+++ packages/utils/src/mergeDeepObject/tests/mergeDeepObject.dot-notation.type-test.ts
@@ -1,0 +1,14 @@
+import { mergeDeepObject } from '../mergeDeepObject';
+
+export const _testDotNotation: {
+    settings: {
+        feature: {
+            enabled: boolean;
+            label: string;
+        };
+    };
+} = mergeDeepObject.withOptions(
+    { dotNotation: true },
+    { 'settings.feature.enabled': true },
+    { settings: { feature: { label: 'enabled' } } },
+);
